backend/pyproject.toml
[project]
name = "samu-backend"
version = "0.1.0"
description = "Samu backend bootstrap"
readme = "README.md"
requires-python = ">=3.11"
keywords = ["fastapi", "samu"]
license = {text = "Proprietary"}
dependencies = [
  "fastapi==0.111.0",
  "uvicorn[standard]==0.29.0",
  "pydantic-settings==2.3.3",
  "SQLAlchemy==2.0.29",
  "alembic==1.13.2",
  "redis==5.0.4",
]

[project.optional-dependencies]
dev = [
  "pytest==8.2.0",
  "httpx==0.27.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]
